Moves utils to 1 plugin, and fixes a few other things
