+++ src/main/resources/META-INF/gradle-plugins/edu.wpi.first.nativeutils.NativeUtils.properties
@@ -1,0 +1,1 @@
+implementation-class=edu.wpi.first.nativeutils.NativeUtils
